requirements.txt
torch>=2.0.0
pettingzoo>=1.24.0
gymnasium>=0.29.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.4.0

# Optional visualization / experiment tracking
# tensorboard>=2.13.0
# matplotlib>=3.7.0
# pygame>=2.5.0
